requirements.txt
requests>=2.31
pandas>=2.1
numpy>=1.24
python-dotenv>=1.0
matplotlib>=3.8
